requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.3.7
python-dotenv==1.0.0
pandas==2.1.3
nltk==3.8.1
textblob==0.17.1
flask==3.0.0
flask-cors==4.0.0
webdriver-manager==4.0.1
lxml==4.9.3
fake-useragent==1.4.0
timeout-decorator==0.5.0
flask>=2.3.0
weasyprint==60.2
markdown==3.8.2
python-docx==1.1.0
